NESSUNO OSI TOCCARE IL MAIN

+++ main.cpp
@@ -4,12 +4,10 @@
 #include <random>
 
 using namespace boidz;
-int main()
-{
+int main() {
   int flock_size{};
   Parameters params{};
-  try
-  {
+  try {
     params.d = read_parameter("Neighbours distance ", 20.f, 100.f);
     params.d_s = read_parameter("Critical distance ", 5.f, 25.f);
     params.s = read_parameter("Separation parameter ", 0.5f, 5.f);
@@ -18,9 +16,7 @@ int main()
     flock_size = read_parameter("Flock size ", 1, 100);
     params.h = 1.;
     params.p = 0.5;
-  }
-  catch (const std::runtime_error &e)
-  {
+  } catch (const std::runtime_error &e) {
     std::cerr << "Errore" << '\n';
     return EXIT_FAILURE;
   }
@@ -32,8 +28,7 @@ int main()
   std::uniform_real_distribution<float> py{0., 200.};
   std::uniform_real_distribution<float> vel{-50., 50.};
 
-  for (int j{0}; j != flock_size; ++j)
-  {
+  for (int j{0}; j != flock_size; ++j) {
     Coords p{px(eng), py(eng)};
     Coords v{vel(eng), vel(eng)};
     SingleBoid b = SingleBoid(p, v, j);
@@ -103,52 +98,49 @@ int main()
 
   float timer = 0.;
   const float delay = 5.;
-  while (window.isOpen() || window2.isOpen() || window3.isOpen())
-  {
+  while (window.isOpen() || window2.isOpen() || window3.isOpen()) {
     sf::Event event;
-    while (window.pollEvent(event))
-    {
-      if (event.type == sf::Event::Closed)
-      {
+    while (window.pollEvent(event)) {
+      if (event.type == sf::Event::Closed) {
         window.close();
       }
     }
-    while (window2.pollEvent(event))
-    {
-      if (event.type == sf::Event::Closed)
-      {
+    while (window2.pollEvent(event)) {
+      if (event.type == sf::Event::Closed) {
         window2.close();
       }
     }
-    while (window3.pollEvent(event))
-    {
-      if (event.type == sf::Event::Closed)
-      {
+    while (window3.pollEvent(event)) {
+      if (event.type == sf::Event::Closed) {
         window3.close();
       }
     }
 
-    if (window.isOpen())
-    {
+    if (window.isOpen()) {
       float dt = clock.restart().asSeconds();
       timer += dt;
       std::vector<Coords> updated_velocities{};
 
-      std::for_each(stormo.flock.begin(), stormo.flock.end(), [&](SingleBoid &b)
-                    {Coords new_velocity = create_velocity_with_rules(params, stormo, b, hunter);
-                     updated_velocities.push_back(new_velocity); });
+      std::for_each(stormo.flock.begin(), stormo.flock.end(),
+                    [&](SingleBoid &b) {
+                      Coords new_velocity =
+                          create_velocity_with_rules(params, stormo, b, hunter);
+                      updated_velocities.push_back(new_velocity);
+                    });
 
       std::for_each(stormo.flock.begin(), stormo.flock.end(),
-                    [&](SingleBoid &b)
-                    {b.update_Velocity_with_rules(updated_velocities[&b - &stormo.flock[0]]);
-                     b.update_Position_in_time(dt); 
-                     b.BorderRestriction(800., 600.);
-                     b.SpeedRestriction(130);
-                     b.update_Shape(); });
+                    [&](SingleBoid &b) {
+                      b.update_Velocity_with_rules(
+                          updated_velocities[static_cast<size_t>(&b - &stormo.flock[0])]);
+                      b.update_Position_in_time(dt);
+                      b.BorderRestriction(800., 600.);
+                      b.SpeedRestriction(130);
+                      b.update_Shape();
+                    });
 
-      if (timer >= delay)
-      {
-        Coords new_hunt_velocity = hunt_the_flock(hunter, stormo, params) + hunt_neighbours(hunter, stormo, params);
+      if (timer >= delay) {
+        Coords new_hunt_velocity = hunt_the_flock(hunter, stormo, params) +
+                                   hunt_neighbours(hunter, stormo, params);
         hunter.update_Velocity_with_rules(new_hunt_velocity);
         hunter.update_Position_in_time(dt);
         hunter.BorderRestriction(800., 600.);
@@ -158,13 +150,12 @@ int main()
 
       Coords mean_speed_for_cycle = mean_velocity(stormo);
       vector_of_mean_speeds.push_back(mean_speed_for_cycle);
-      if (vector_of_mean_speeds.size() > max_points)
-      {
+      if (vector_of_mean_speeds.size() > max_points) {
         vector_of_mean_speeds.erase(
             vector_of_mean_speeds.begin(),
             vector_of_mean_speeds.begin() +
-                (static_cast<float>(vector_of_mean_speeds.size()) - static_cast<float>(
-                 max_points)));
+                static_cast<long int>(
+                    (vector_of_mean_speeds.size() - max_points)));
       }
       float standard_deviation_velocity = standard_deviation(stormo.flock);
       vector_of_standard_deviations_v.push_back(standard_deviation_velocity);
@@ -172,43 +163,43 @@ int main()
       float mean_distance_boidz = mean_distance(stormo);
       vector_of_mean_distances.push_back(mean_distance_boidz);
       auto vector_of_distances = get_vector_of_distances(stormo);
-      float standard_deviation_distance = standard_deviation(vector_of_distances);
+      float standard_deviation_distance =
+          standard_deviation(vector_of_distances);
       vector_of_standard_deviations_d.push_back(standard_deviation_distance);
-      if (vector_of_mean_distances.size() > max_points)
-      {
+      if (vector_of_mean_distances.size() > max_points) {
         vector_of_mean_distances.erase(
             vector_of_mean_distances.begin(),
             vector_of_mean_distances.begin() +
-                (static_cast<float>(vector_of_mean_distances.size()) - static_cast<float>(
-                 max_points))); // toglie i punti extra dal grafico
+                static_cast<long int>(
+                    (vector_of_mean_distances.size() -
+                     max_points))); // toglie i punti extra dal grafico
       }
-      if (vector_of_standard_deviations_v.size() > max_points)
-      {
+      if (vector_of_standard_deviations_v.size() > max_points) {
         vector_of_standard_deviations_v.erase(
             vector_of_standard_deviations_v.begin(),
             vector_of_standard_deviations_v.begin() +
-                (static_cast<float>(vector_of_standard_deviations_v.size()) - static_cast<float>(
-                 max_points))); // toglie i punti extra dal grafico
+                static_cast<long int>(
+                    (vector_of_standard_deviations_v.size() -
+                     max_points))); // toglie i punti extra dal grafico
       }
-      if (vector_of_standard_deviations_d.size() > max_points){
+      if (vector_of_standard_deviations_d.size() > max_points) {
         vector_of_standard_deviations_d.erase(
             vector_of_standard_deviations_d.begin(),
             vector_of_standard_deviations_d.begin() +
-                (static_cast<float>(vector_of_standard_deviations_d.size()) - static_cast<float>(
-                 max_points))); // toglie i punti extra dal grafico
+                static_cast<long int>(
+                    (vector_of_standard_deviations_d.size() -
+                     max_points))); // toglie i punti extra dal grafico
       }
 
       window.clear(sf::Color(15, 17, 26));
-      for (auto &b : stormo.flock)
-      {
+      for (auto &b : stormo.flock) {
         b.draw(window);
       }
       hunter.draw(window);
       window.display();
     }
 
-    if (window2.isOpen())
-    {
+    if (window2.isOpen()) {
       window2.clear();
       window2.draw(axis);
 
@@ -217,31 +208,27 @@ int main()
       float graph_right = 450.0;
       float graph_width = graph_right - graph_left;
       float x_step;
-      if (visible_count > 1)
-      {
+      if (visible_count > 1) {
         x_step = graph_width / static_cast<float>((visible_count - 1));
-      }
-      else
-      {
+      } else {
         x_step = 0.0;
       }
 
-      for (long unsigned int i{0}; i < visible_count; ++i)
-      {
-        float output_speed = std::sqrt(vector_of_mean_speeds[i].x * vector_of_mean_speeds[i].x + 
-          vector_of_mean_speeds[i].y * vector_of_mean_speeds[i].y);
-        float x = graph_left + static_cast<float>(x_step * i);
+      for (long unsigned int i{0}; i < visible_count; ++i) {
+        float output_speed =
+            std::sqrt(vector_of_mean_speeds[i].x * vector_of_mean_speeds[i].x +
+                      vector_of_mean_speeds[i].y * vector_of_mean_speeds[i].y);
+        float x = graph_left + (x_step * static_cast<float>(i));
         float y = 550.0f - output_speed;
-        if (y < 50.0f)
-        {
+        if (y < 50.0f) {
           y = 50.0f;
         }
-        if (y > 550.0f)
-        {
+        if (y > 550.0f) {
           y = 550.0f;
         }
-        label.setString("Mean speed = " + std::to_string(output_speed) + '\n' +
-                        "Error = " + std::to_string(vector_of_standard_deviations_v[i]));
+        label.setString(
+            "Mean speed = " + std::to_string(output_speed) + '\n' +
+            "Error = " + std::to_string(vector_of_standard_deviations_v[i]));
         point.setPosition(x, y);
         window2.draw(point);
       }
@@ -250,8 +237,7 @@ int main()
       window2.display();
     }
 
-    if (window3.isOpen())
-    {
+    if (window3.isOpen()) {
       window3.clear();
       window3.draw(axis);
 
@@ -261,30 +247,25 @@ int main()
       float graph_right = 450.0;
       float graph_width = graph_right - graph_left;
       float x_step;
-      if (visible_count2 > 1)
-      {
+      if (visible_count2 > 1) {
         x_step = graph_width / static_cast<float>((visible_count2 - 1));
-      }
-      else
-      {
+      } else {
         x_step = 0.0;
       }
       for (auto first = vector_of_mean_distances.begin(),
                 last = vector_of_mean_distances.end();
-           first != last && i < visible_count2; ++first, ++i)
-      {
-        float x = graph_left + static_cast<float>(x_step * i);
+           first != last && i < visible_count2; ++first, ++i) {
+        float x = graph_left + (x_step * static_cast<float>(i));
         float y = 550.0f - *first;
-        if (y < 50.0f)
-        {
+        if (y < 50.0f) {
           y = 50.0f;
         }
-        if (y > 550.0f)
-        {
+        if (y > 550.0f) {
           y = 550.0f;
         }
-        label2.setString("Mean distance =" + std::to_string(*first) + '\n' +
-                         "Error = " + std::to_string(vector_of_standard_deviations_d[i]));
+        label2.setString(
+            "Mean distance =" + std::to_string(*first) + '\n' +
+            "Error = " + std::to_string(vector_of_standard_deviations_d[i]));
         dot.setPosition(x, y);
         window3.draw(dot);
       }
